backend/requirements.txt
fastapi==0.132.0
uvicorn==0.41.0
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
python-docx==1.2.0
reportlab==4.4.10
pandas==3.0.1
PyPDF2==3.0.1
